refactor(Model): пространство имен Lab1Model переименовано в Model
