+++ Model/Resistor.cs
@@ -1,7 +1,7 @@
 ﻿using System.Numerics;
 
 
-namespace Lab1Model
+namespace Model
 {
 	namespace PassiveComponents
 	{
