+++ View/Forms/MainForm.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Numerics;
 using System.Windows.Forms;
 
-using Lab1Model;
+using Model;
 using PositiveDoubleTextBoxLib;
